feat: add products API endpoints
This patch adds a working products API endpoint and a (slightly) broken
endpoint that returns nothing if the user does not explicitly specify a
limit and skip in their URL search parameters. This is due to a somewhat
subtle bug where `Number(null)` evaluates to `0` instead of `NaN`.

I will likely ask interviewees to debug and fix this issue. It took me a
few minutes only. I added a `console.log` and quickly realized my
mistake. I will see what they do. They should be able to solve this
quickly as well.